WebApp/backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
chromadb
pypdf
python-dotenv
markdown
